stutter_correction/requirements.txt
faster-whisper
transformers
groq
streamlit
torch
torchaudio
librosa
numpy
scipy
soundfile
pydantic
python-dotenv
pytest
